ENH: Add static methods basis and cast to the polynomial classes.
The new basis method is a convenient way to return an instance of the basis
function of given degree for the class. It is intended mostly for
pedagogical purposes.

The new cast method provides an alternate way to convert an instance of one
polynomial class to another. It complements the convert instance method.