STY, BUG: Cleanup some macros and move a NULL check.
Some macros are changed into statements with the {...}do while(0)
construction, while a few others have had the trailing semicolon
removed. This makes the macro usage a bit safer and more consistent,
but it is possible some third party code will be affected.

A NULL check was moved ahead of the spot where the variable is
first used.