DEP: Fix up handling of the O4/O8 deprecation
First issue was that the deprecation warning was still
being raised even when an error was being raised (like for 'O3').

Second issue was that unpickling of legacy pickles which have 'O8'
or 'O4' should not raise the deprecation warning. To handle this
required introduction of a global variable, which has been named
and commented to make it clear this was a necessary evil for this
specific case, not a standard NumPy practice.