BUG: ticket #1578, Fix python-debug warning for python >= 2.7.
In Python >= 2.7 the memory management of unicode objects changed from
PyMem_* to PyObject_*. Numpy had its own modified copy of some of the
Python code for handling unicode objects and it needed to be updated
to deal with the change. Thanks to Joseph Miessner for tracking this down
and finding the fix.