TST: filter ImportWarnings in NoseTester.
Warnings show up when a directory with the same name as a Python file or
compiled extension is seen which doesn't have an __init__.py file in it.  This
situation is very common, for example in SciPy where many extensions are
created from source files located under a directory with the same name.

This filter is located within a context manager, so only filters when running
tests.