BUG: Bubble errors from _tuple_of_integers.
Refactors _tuple_of_integers a little bit in order to propagate error
state. Zero indicates failure of the test (given tuple is not a tuple of
integers), whereas non-zero indicates either success (positive) or
an exception during the warning (negative). This allows us to use
(ret = _tuple_of_integers(...)) as an if condition, and then put the
error handling inside the if body.