2to3: Skip buffer fixer.
The buffer object is replaced by memoryview in Python >= 3. The memory
view object has also been backported to Python 2.7. However, the only
use of `buffer1/1memoryview1 is in `numpy/core/tests/test_unicode.py`
and there it is already version dependent: `memoryview` is used if the
Python version is >= 3 and `buffer` is used otherwise.

Closes #3043.